requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
pydantic==2.11.3
pydantic-settings==2.9.1
grpcio==1.72.0
grpcio-tools==1.72.0
protobuf==6.30.0
structlog==25.4.0
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-instrumentation-grpc==0.54b0
opentelemetry-exporter-otlp==1.33.0
lingua-language-detector==2.1.0
nltk==3.9.1
fastembed==0.6.1
qdrant-client==1.14.2
transformers==4.52.4
torch==2.7.0
sentencepiece==0.2.0
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.1.1
httpx==0.28.1
ruff==0.11.12
